Update README.md
## Correction of Spelling and SDK Requirements

### Description
In this contribution, I've made the following changes to the documentation:

1. Corrected a spelling error in the README.
2. Updated the `minSdkVersion` and `targetSdkVersion` requirements in the README.

### Spelling Correction
I corrected a few minor spelling errors in the project's README. 

### SDK Requirements Update
I updated the `minSdkVersion` to 21 and the `targetSdkVersion` to 28 in the README. This change ensures compatibility with modern Android platforms and takes advantage of the latest features. 

### Impact
These changes improve the overall quality and accuracy of the project's documentation. Users and developers will benefit from clearer and more precise information.

### Testing
I tested the updated SDK requirements on an Android emulator to ensure their correctness.